+++ sparta.go
@@ -631,6 +631,9 @@ type LambdaAWSInfo struct {
 	customResources []*customResourceInfo
 	// Cached lambda name s.t. we only compute it once
 	cachedLambdaFunctionName string
+
+	// deprecation notices
+	deprecationNotices []string
 }
 
 // lambdaFunctionName returns the internal
@@ -639,8 +642,11 @@ func (info *LambdaAWSInfo) lambdaFunctionName() string {
 	if info.cachedLambdaFunctionName != "" {
 		return info.cachedLambdaFunctionName
 	}
-	lambdaFuncName := info.userSuppliedFunctionName
-	if nil != info.Options &&
+	var lambdaFuncName string
+
+	if info.userSuppliedFunctionName != "" {
+		lambdaFuncName = info.userSuppliedFunctionName
+	} else if nil != info.Options &&
 		nil != info.Options.SpartaOptions &&
 		"" != info.Options.SpartaOptions.Name {
 		lambdaFuncName = info.Options.SpartaOptions.Name
@@ -648,6 +654,8 @@ func (info *LambdaAWSInfo) lambdaFunctionName() string {
 		// Using the default name, let's at least remove the
 		// first prefix, since that's the SCM provider and
 		// doesn't provide a lot of value...
+		// Why are we overriding the supplied name?
+
 		if info.handlerSymbol != nil {
 			lambdaPtr := runtime.FuncForPC(reflect.ValueOf(info.handlerSymbol).Pointer())
 			lambdaFuncName = lambdaPtr.Name()
@@ -966,7 +974,15 @@ func (info *LambdaAWSInfo) export(serviceName string,
 		context,
 		logger)
 
-	return decoratorErr
+	if decoratorErr != nil {
+		return decoratorErr
+	}
+
+	// Log any deprecation notices
+	for _, eachDeprecation := range info.deprecationNotices {
+		logger.Warn(eachDeprecation)
+	}
+	return nil
 }
 
 //
@@ -1100,11 +1116,19 @@ Supported lambdaHandler signatures:
 • func (context.Context, TIn) (TOut, error)
 */
 
-// HandleAWSLambda registers lambdaHandler with the given functionName
-// using the default lambdaFunctionOptions
-func HandleAWSLambda(functionName string,
+// NewAWSLambda is the creation function that replaces HandleAWSLambda. It returns
+// a *LambdaAWSInfo pointer to the struct representing the AWS lambda target. It's a
+// go-friendly signature for creating a lambda function
+func NewAWSLambda(functionName string,
 	lambdaHandler interface{},
-	roleNameOrIAMRoleDefinition interface{}) *LambdaAWSInfo {
+	roleNameOrIAMRoleDefinition interface{}) (*LambdaAWSInfo, error) {
+
+	if functionName == "" {
+		return nil, errors.Errorf("AWS Lambda function name must not be empty")
+	}
+	if lambdaHandler == nil {
+		return nil, errors.Errorf("AWS Lambda function handler must not be nil")
+	}
 
 	lambda := &LambdaAWSInfo{
 		userSuppliedFunctionName: functionName,
@@ -1112,17 +1136,31 @@ func HandleAWSLambda(functionName string,
 		Options:                  defaultLambdaFunctionOptions(),
 		Permissions:              make([]LambdaPermissionExporter, 0),
 		EventSourceMappings:      make([]*EventSourceMapping, 0),
+		deprecationNotices:       make([]string, 0),
 	}
 
 	switch v := roleNameOrIAMRoleDefinition.(type) {
 	case string:
-		lambda.RoleName = roleNameOrIAMRoleDefinition.(string)
+		lambda.RoleName = v
 	case IAMRoleDefinition:
-		definition := roleNameOrIAMRoleDefinition.(IAMRoleDefinition)
+		definition := v
 		lambda.RoleDefinition = &definition
 	default:
-		panic(fmt.Sprintf("Unsupported IAM Role type: %s", v))
+		return nil, errors.Errorf("AWS Lambda function IAM role must not be empty")
 	}
+	return lambda, nil
+}
+
+// HandleAWSLambda is deprecated in favor of NewAWSLambda(...)
+func HandleAWSLambda(functionName string,
+	lambdaHandler interface{},
+	roleNameOrIAMRoleDefinition interface{}) *LambdaAWSInfo {
+
+	lambda, lambdaErr := NewAWSLambda(functionName, lambdaHandler, roleNameOrIAMRoleDefinition)
+	if lambdaErr != nil {
+		panic(lambdaErr)
+	}
+	lambda.deprecationNotices = append(lambda.deprecationNotices, "sparta.HandleAWSLambda is deprecated starting with v1.6.0. Prefer sparta.NewAWSLambda(...)")
 	return lambda
 }